# [WIP] Speed up geojson output