Trac issue: https://trac.osgeo.org/postgis/ticket/4615

Printing the doubles directly into the output buffer doubles the tps with big inputs:

Before 02e62ee:
```
number of clients: 1
number of threads: 1
duration: 30 s
number of transactions actually processed: 31
latency average = 975.931 ms
tps = 1.024663 (including connections establishing)
tps = 1.024788 (excluding connections establishing)
```

After 02e62ee:
```
number of clients: 1
number of threads: 1
duration: 30 s
number of transactions actually processed: 58
latency average = 518.217 ms
tps = 1.929693 (including connections establishing)
tps = 1.929935 (excluding connections establishing)
```

After the introduction of Ryu (https://trac.osgeo.org/postgis/ticket/4543) there are new rough corners, so I plan to keep polishing for a while before merging.