@Algunenano 2020-01-24
Some notes about other things I've tested:
- I've removed the need to calculate the length of the final string (aka, use `cstring_to_text_with_len` instead of `cstring_to_text`) by passing the size to `lwgeom_to_geojson` as an output value: Initially I introduced a bug and the size returned was the max calculated size, not the actual output size which made 6% slower and left me surprised; after fixing it: 1% faster. Not worth the trouble.

- As mentioned in https://trac.osgeo.org/postgis/ticket/4623, I've tried to create a new liblwgeom type that contains both the size and the data, that is, a `varlena` compatible type so there is no need to transform the data from liblwgeom to Postgresql: 7% improvement, might be worth the trouble even if it's just for geojson (but I expect the impact on other functions to be much higher since the time spent on the function bodies is smaller, e.g St_AsBinary): https://github.com/Algunenano/postgis/commit/6f86f18335694d97e9c48ceaa93073ae82977293


After the previous change (the varlena type) the current process with big geometries show an ussage like this:
```
+   98.53%     0.01%             2  postgres  postgis-3.so            [.] LWGEOM_asGeoJson                                                                                                                                                                  ▒
+   94.78%     0.02%            45  postgres  postgis-3.so            [.] lwgeom_to_geojson                                                                                                                                                                 ▒
+   94.63%     0.04%            65  postgres  postgis-3.so            [.] asgeojson_multipolygon_buf                                                                                                                                                        ▒
+   94.41%     4.05%          4720  postgres  postgis-3.so            [.] pointArray_to_geojson                                                                                                                                                             ▒
+   90.51%     7.85%          8992  postgres  postgis-3.so            [.] lwprint_double                                                                                                                                                                    ▒
+   82.84%    80.38%         90489  postgres  postgis-3.so            [.] d2fixed_buffered_n
```

So technically, between the 98.53% for LWGEOM_asGeoJson and the time spent printing doubles (which should be the high as possible), that is 82.84%, there is only 16% of time being spent, which is low. I might still have a look to see if anything can be improved, but I might end up pushing only this commit here and the varlena change as part of a different PR implementing it for other functions.